backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speaking-practice-backend"
version = "0.1.0"
description = "AI-powered speaking practice application backend"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "openai>=1.35.0",
    "pydub>=0.25.1",
    "kokoro>=0.9.4",
    "soundfile",
    "numpy<2.0; sys_platform != 'darwin'",
    "numpy>=2.2.5; sys_platform == 'darwin'",
    "torch; sys_platform != 'darwin'",
    "scipy; sys_platform != 'darwin'",
    "typer",
    "parakeet-mlx; sys_platform == 'darwin'",
    "nemo_toolkit[asr]; sys_platform != 'darwin'",
    "cuda-python>=12.3; sys_platform != 'darwin'",
    "hf-transfer>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    # Linting and formatting
    "flake8",
    "flake8-docstrings",
    "flake8-bugbear",
    "flake8-comprehensions",
    "black",
    "isort",
    "mypy",
    "pre-commit",
]

[project.scripts]
dev = "app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "flake8-bugbear>=25.11.29",
    "flake8-comprehensions>=3.17.0",
    "flake8-docstrings>=1.7.0",
    "flake8-functions>=0.0.8",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pip>=25.3",
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
]

# Black configuration
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort configuration
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]
known_third_party = ["fastapi", "pydantic", "pytest", "httpx", "openai", "uvicorn"]

# MyPy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "parakeet_mlx.*",
    "nemo.*",
    "torch.*",
    "kokoro.*",
    "soundfile.*",
    "pydub.*",
]
ignore_missing_imports = true

# Flake8 configuration
[tool.flake8]
max-line-length = 88
max-doc-length = 88
max-complexity = 10
max-file-length = 300
max-function-length = 50
extend-ignore = [
    "E203",  # whitespace before ':' (conflicts with black)
    "W503",  # line break before binary operator (conflicts with black)
    "E501",  # line too long (handled by black)
]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".eggs",
    "*.egg",
    "build",
    "dist",
    ".mypy_cache",
    ".pytest_cache",
]
per-file-ignores = [
    "__init__.py:F401",  # imported but unused
    "test_*.py:F401,F811",  # test files can have unused imports and redefined functions
]
